// src/web-app/web-request-handlers/github-webhook/transformGithubPayloadToAppEvent/GitHubAPI.contracts.d.ts
export type GitHubPullRequestEditedNotification = GitHubPullRequestNotificationBasicPayload & {
    readonly action: "edited";
    readonly changes: {
        readonly body?: {
            readonly from: string;
        };
        readonly title?: {
            from: string;
        };
        readonly base?: {
            readonly ref: {
                from: string;
            };
            readonly sha: {
                readonly from: string;
            };
        };
    };
};
// ...
export type GitHubPullRequestNotificationBasicPayload = {
    readonly pull_request: GitHubPullRequestPayload;
    readonly repository: {
        readonly name: string;
    };
    readonly sender: GitHubUserPayload;
    readonly organization: { readonly login: string; id: number };
};

export type GitHubPullRequestCommentPayload = {
    readonly html_url: string;
    readonly id: number;
    readonly path?: string;
    readonly user: GitHubUserPayload;
    readonly body: string;
    readonly updated_at: string;
    readonly in_reply_to_id?: number;
};

export type GitHubPullRequestPayload = {
    readonly html_url: string;
    readonly number: number;
    readonly state: "open" | "closed";
    readonly locked: boolean;
    readonly title: string;
    readonly user: GitHubUserPayload;
    readonly body: string | null;
    readonly created_at: string;
    readonly updated_at: string;
    readonly closed_at: string | null;
    readonly merged_at: string | null;
    readonly merge_commit_sha: string | null;
    readonly assignee: string | null;
    readonly assignees: GitHubUserPayload[];
    readonly requested_reviewers: GitHubUserPayload[];
    readonly requested_teams: GitHubTeamPayload[];
    readonly draft: boolean;
    readonly head: GitHubRefPayload;
    readonly base: GitHubRefPayload;
    readonly merged: boolean;
    readonly review_comments: number;
};
export type GitHubTeamPayload = {
    members_url: string;
};
export type GitHubUserPayload = {
    readonly login: string;
    readonly type: "User" | "Mannequin" | "Bot";
    readonly html_url: string;
};
// ...
export type GitHubPullRequestEventType =
    | "pull_request"
    | "issue_comment"
    | "pull_request_review_comment"
    | "pull_request_review_thread"
    | "pull_request_review";